Posts/Users foreign

+++ cms/database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('role_id')->index()->unsigned()->nullable();
+            $table->bigInteger('role_id')->index()->unsigned()->nullable();
             $table->boolean('is_active')->default(true);
             $table->rememberToken();
             $table->timestamps();
